requirements.txt
requests>=2.31.0
google-generativeai>=0.3.0
PyYAML>=6.0
colorama>=0.4.6